+++ dicelib/tractogram/__init__.py
@@ -1,4 +1,0 @@
-# __all__ = ['clustering', 'processing']
-#from dicelib.tractogram import clustering
-#from dicelib.tractogram import dummy
-# from dicelib.tractogram import processing