+++ dicelib/tractogram/processing_c.cpp
@@ -12,25 +12,41 @@ int do_spline_smoothing( float* ptr_npaFiberI, int nP, float* ptr_npaFiberO, flo
     std::vector<float>          polyline_simplified;
     std::vector<Vector<float>>  CPs;
     Catmull                     FIBER;
+    int                         n;
 
-    // simplify input polyline down to nP*ratio points
-    psimpl::simplify_douglas_peucker_n<3>( ptr_npaFiberI, ptr_npaFiberI+3*nP, nP*ratio, std::back_inserter(polyline_simplified) );
-
-    CPs.resize( polyline_simplified.size()/3 );
-    for( int j=0,index=0; j < polyline_simplified.size(); j=j+3 )
-        CPs[index++].Set( polyline_simplified[j], polyline_simplified[j+1], polyline_simplified[j+2] );
-
-    // perform interpolation
-    FIBER.set( CPs );
-    FIBER.eval( FIBER.L/segment_len );
-    FIBER.arcLengthReparametrization( segment_len );
-
-    // copy coordinates of the smoothed streamline back to python
-    for( int j=0; j < FIBER.P.size(); j++ )
+    if ( nP<=2 )
     {
-        *(ptr_npaFiberO++) = FIBER.P[j].x;
-        *(ptr_npaFiberO++) = FIBER.P[j].y;
-        *(ptr_npaFiberO++) = FIBER.P[j].z;
+        // if input streamline has less than 2 points, just copy input to output
+        for( int j=0; j<3*nP; j++ )
+            *(ptr_npaFiberO++) = *(ptr_npaFiberI++);
+        return nP;
     }
-    return FIBER.P.size();
+    else
+    {
+        // check that at least 3 points are considered
+        n = nP*ratio;
+        if ( n<3 )
+            n = 3;
+
+        // simplify input polyline down to n points
+        psimpl::simplify_douglas_peucker_n<3>( ptr_npaFiberI, ptr_npaFiberI+3*nP, n, std::back_inserter(polyline_simplified) );
+
+        CPs.resize( polyline_simplified.size()/3 );
+        for( int j=0,index=0; j < polyline_simplified.size(); j=j+3 )
+            CPs[index++].Set( polyline_simplified[j], polyline_simplified[j+1], polyline_simplified[j+2] );
+
+        // perform interpolation
+        FIBER.set( CPs );
+        FIBER.eval( FIBER.L/segment_len );
+        FIBER.arcLengthReparametrization( segment_len );
+
+        // copy coordinates of the smoothed streamline back to python
+        for( int j=0; j<FIBER.P.size(); j++ )
+        {
+            *(ptr_npaFiberO++) = FIBER.P[j].x;
+            *(ptr_npaFiberO++) = FIBER.P[j].y;
+            *(ptr_npaFiberO++) = FIBER.P[j].z;
+        }
+        return FIBER.P.size();
+    }
 }